+++ sable_box3d/src/main/java/dev/ryanhcode/sable/physics/impl/box3d/box/Box3dBoxHandle.java
@@ -7,7 +7,7 @@ import dev.ryanhcode.sable.physics.impl.box3d.Box3dJNI;
 import org.joml.Quaterniondc;
 import org.joml.Vector3dc;
 
-public record Box3dBoxHandle(long world, long body, float[] poseCache) implements BoxHandle {
+public record Box3dBoxHandle(long world, long body, double[] poseCache) implements BoxHandle {
 
     public static Box3dBoxHandle create(final long worldHandle, final Pose3dc pose, final Vector3dc halfExtents, final float mass) {
         final Vector3dc pos = pose.position();
@@ -15,7 +15,7 @@ public record Box3dBoxHandle(long world, long body, float[] poseCache) implement
 
         final int id = Box3dJNI.nextBodyID();
         final long bodyHandle = Box3dJNI.createBox(worldHandle, mass, (float) halfExtents.x(), (float) halfExtents.y(), (float) halfExtents.z(), new float[]{(float) pos.x(), (float) pos.y(), (float) pos.z(), (float) rot.x(), (float) rot.y(), (float) rot.z(), (float) rot.w()});
-        return new Box3dBoxHandle(worldHandle, bodyHandle, new float[7]);
+        return new Box3dBoxHandle(worldHandle, bodyHandle, new double[7]);
     }
 
     /**
